WIP: token keeper, query + types
