+++ hs-abci-examples/nameservice/package.yaml
@@ -79,6 +79,9 @@ library:
   - Nameservice.Server
   - Nameservice.Handlers
   - Nameservice.Modules.Token
+  - Nameservice.Modules.Token.Types
+  - Nameservice.Modules.Token.Keeper
+  - Nameservice.Modules.Token.Query
   - Nameservice.Modules.Nameservice.Messages
   - Nameservice.Modules.Nameservice.Types
   - Nameservice.Modules.Nameservice.Keeper
