+++ hs-abci-examples/nameservice/src/Nameservice/Modules/Token/Keeper.hs
@@ -1,0 +1,6 @@
+module Nameservice.Modules.Token.Keeper where
+
+import qualified Tendermint.SDK.Store            as Store
+
+storeKey :: Store.StoreKey "token"
+storeKey = Store.StoreKey "token"
